// frontend/src/app/models/user.ts
export class User {
    firstname: string;
    surname: string;
    username: string;
    password: string;
    email: string;
    phone: string;
    agency: string;
    license: string;
    birthday: string;
    type: number;//1 kupac 2 adviser 3 admin
    city: string;
    imageFile: string;
    favorites: number[];
}